Getting rid of CCEffect_Private.h
