+++ cocos2d-ext/CCEffects/CCEffect.m
@@ -7,7 +7,6 @@
 //
 
 #import "CCEffect.h"
-#import "CCEffect_Private.h"
 #import "CCTexture.h"
 
 NSString * const CCShaderUniformPreviousPassTexture = @"cc_PreviousPassTexture";
