+++ cocos2d-ext/CCEffects/CCEffectBloom.m
@@ -42,7 +42,7 @@
 
 #import "CCEffectBloom.h"
 #import "CCEFfectUtils.h"
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCRenderer.h"
 #import "CCTexture.h"
