+++ cocos2d-ext/CCEffects/CCEffectBlur.m
@@ -41,7 +41,7 @@
 
 #import "CCEffectBlur.h"
 #import "CCEffectUtils.h"
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCTexture.h"
