+++ cocos2d-ext/CCEffects/CCEffectBrightness.m
@@ -7,7 +7,7 @@
 //
 
 #import "CCEffectBrightness.h"
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCRenderer.h"
 #import "CCTexture.h"
