+++ cocos2d-ext/CCEffects/CCEffectColorChannelOffset.m
@@ -7,7 +7,7 @@
 //
 
 #import "CCEffectColorChannelOffset.h"
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCRenderer.h"
 #import "CCTexture.h"
