+++ cocos2d-ext/CCEffects/CCEffectContrast.m
@@ -7,7 +7,7 @@
 //
 
 #import "CCEffectContrast.h"
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCRenderer.h"
 #import "CCTexture.h"
