+++ cocos2d-ext/CCEffects/CCEffectHue.m
@@ -7,7 +7,7 @@
 
 
 #import "CCEffectHue.h"
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCRenderer.h"
 #import "CCTexture.h"
