+++ cocos2d-ext/CCEffects/CCEffectLighting.m
@@ -18,7 +18,7 @@
 #import "CCSpriteFrame.h"
 #import "CCTexture.h"
 
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCSprite.h"
