+++ cocos2d-ext/CCEffects/CCEffectPixellate.m
@@ -41,7 +41,7 @@
 
 
 #import "CCEffectPixellate.h"
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCRenderer.h"
 #import "CCTexture.h"
