+++ cocos2d-ext/CCEffects/CCEffectRefraction.m
@@ -14,7 +14,7 @@
 #import "CCSpriteFrame.h"
 #import "CCTexture.h"
 
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCSprite.h"
