+++ cocos2d-ext/CCEffects/CCEffectRenderer.m
@@ -13,8 +13,7 @@
 #import "CCEffectStack.h"
 #import "CCEffectUtils.h"
 #import "CCTexture.h"
-
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCRenderer_Private.h"
 #import "CCSprite.h"
 #import "CCTexture.h"
