+++ cocos2d-ext/CCEffects/CCEffectSaturation.m
@@ -41,7 +41,7 @@
 
 
 #import "CCEffectSaturation.h"
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCRenderer.h"
 #import "CCTexture.h"
