+++ cocos2d-ext/CCEffects/CCEffectStereo.m
@@ -10,7 +10,7 @@
 
 #if CC_EFFECTS_EXPERIMENTAL
 
-#import "CCEffect_Private.h"
+#import "CCEffect.h"
 #import "CCRenderer.h"
 #import "CCTexture.h"
